Modularize variable selector component
